requirements.txt
alembic==1.11.1
flask==2.3.2
flask_sqlalchemy==3.0.3
flask_login==0.6.2
python-dotenv==1.0.0
psycopg2==2.9.6
Flask-Migrate==4.0.4
opencv-python==4.7.0.72
matplotlib==3.7.1
numpy==1.24.3
Pillow==9.5.0
pdf2image==1.16.3
werkzeug==2.3.5
scikit-image==0.21.0